+++ src/inmanta_module_factory/inmanta/attribute.py
@@ -20,6 +20,9 @@ from typing import Optional, Union
 
 from typing_extensions import Literal
 
+from inmanta_module_factory.inmanta import entity as inmanta_entity
+from inmanta_module_factory.inmanta import entity_field
+
 InmantaPrimitiveType = Literal["string", "int", "float", "number", "bool"]
 
 
@@ -34,7 +37,7 @@ class InmantaPrimitiveList:
 InmantaAttributeType = Union[Literal["dict", "any"], InmantaPrimitiveType, InmantaPrimitiveList]
 
 
-class Attribute:
+class Attribute(entity_field.EntityField):
     def __init__(
         self,
         name: str,
@@ -42,6 +45,7 @@ class Attribute:
         optional: bool = False,
         default: Optional[str] = None,
         description: Optional[str] = None,
+        entity: Optional["inmanta_entity.Entity"] = None,
     ) -> None:
         """
         :param name: The name of the attribute
@@ -49,8 +53,9 @@ class Attribute:
         :param optional: Whether this attribute is optional or not
         :param default: Whether this attribute has a default value or not
         :param description: A description of the attribute to add in the docstring
+        :param entity: The entity this attribute is a part of
         """
-        self.name = name
+        entity_field.EntityField.__init__(self, name, entity)
         self._inmanta_type = inmanta_type
         self.optional = optional
         self.default = default