+++ tests/test_simple_module.py
@@ -54,27 +54,29 @@ def test_basic_module(project: Project) -> None:
     entity = Entity(
         "Test",
         path=[module.name],
-        attributes=[
+        fields=[
             Attribute(
                 name="test",
                 inmanta_type=InmantaPrimitiveList("string"),
                 default="[]",
                 description="This is a test list attribute",
             ),
-            Attribute(
-                name="test1",
-                inmanta_type="string",
-                description="This is a test attribute",
-            ),
         ],
         description="This is a test entity",
+    )
+
+    index_attribute = Attribute(
+        name="test1",
+        inmanta_type="string",
+        description="This is a test attribute",
+        entity=entity,
     )
 
     implementation = Implementation(
         name="test",
         path=[module.name],
         entity=entity,
-        content=f"std::print(self.{entity.attributes[1].name})",
+        content=f"std::print(self.{index_attribute.name})",
         description="This is a test implementation",
     )
 
@@ -87,7 +89,7 @@ def test_basic_module(project: Project) -> None:
     index = Index(
         path=[module.name],
         entity=entity,
-        attributes=entity.attributes[1:2],
+        fields=[index_attribute],
         description="This is a test index",
     )
 
@@ -95,12 +97,12 @@ def test_basic_module(project: Project) -> None:
         name="tests",
         path=[module.name],
         entity=entity,
-        arity=(0, None),
+        cardinality=(0, None),
         peer=EntityRelation(
             name="",
             path=[module.name],
             entity=entity,
-            arity=(0, 0),
+            cardinality=(0, 0),
         ),
     )